Renames GitCommands command to QuickWizard
Updates telemetry source to 'quick-wizard'
