+++ src/commands/git/worktree.ts
@@ -40,7 +40,6 @@ import { configuration } from '../../system/vscode/configuration';
 import { isDescendant } from '../../system/vscode/path';
 import { getWorkspaceFriendlyPath, openWorkspace, revealInFileExplorer } from '../../system/vscode/utils';
 import type { ViewsWithRepositoryFolders } from '../../views/viewBase';
-import { getSteps } from '../gitCommands.utils';
 import type {
 	AsyncStepResultGenerator,
 	CustomStep,
@@ -70,6 +69,7 @@ import {
 	pickWorktreesStep,
 	pickWorktreeStep,
 } from '../quickCommand.steps';
+import { getSteps } from '../quickWizard.utils';
 
 interface Context {
 	repos: Repository[];
