+++ src/constants.telemetry.ts
@@ -12,7 +12,6 @@ export type Sources =
 	| 'cloud-patches'
 	| 'commandPalette'
 	| 'deeplink'
-	| 'git-commands'
 	| 'graph'
 	| 'home'
 	| 'inspect'
@@ -24,6 +23,7 @@ export type Sources =
 	| 'notification'
 	| 'patchDetails'
 	| 'prompt'
+	| 'quick-wizard'
 	| 'remoteProvider'
 	| 'settings'
 	| 'timeline'
